#19: more documentation and some light refactor
